modules/tools/browsers: init
This rework the abstraction module for Firefox into a generic
abstraction for both major browsers: Chromium and Firefox.

This enables developers to let user choose between Firefox and Chromium
and configure them uniformly.

Obviously, abstracting all options is impossible, so a developer should
always consider using the fine-grained API of a browser if needed.

Signed-off-by: Ryan Lahfa <ryan.lahfa.ext@numerique.gouv.fr>