pkgs/plasma/mkPlasmaLookAndFeelPackage: init
This enables the construction of ~arbitrary ad-hoc Look and Feel
packages.

Plasma 6 has very few documentation on how to achieve this, it's very
recommended to read carefully:

https://develop.kde.org/docs/plasma/#global-theme
https://develop.kde.org/docs/plasma/scripting/

and read a bunch of examples.

We provide only bare minimum: wallpaper, application launcher icons and
everything that can be configured via KConfig reasonably.

Splash screens may be configured but this has not been tested.
SDDM may be configured but this has not been tested.

Signed-off-by: Ryan Lahfa <ryan.lahfa.ext@numerique.gouv.fr>